Feat/account

See merge request s10-webmobile2-sub2/S10P12A209!170

+++ BE/src/main/java/com/ssafy/matdongsan/domain/account/model/PersonTag.java
@@ -25,7 +25,7 @@ public class PersonTag extends BaseEntity {
     @Column(name = "person_tag_name", nullable = false, length = 100)
     private String name;
 
-    private int birthYear;
+    private short birthYear;
 
     @ManyToOne(fetch = LAZY)
     @JoinColumn(name = "account_id")
@@ -36,7 +36,7 @@ public class PersonTag extends BaseEntity {
 
 
     @Builder
-    public PersonTag(String name, int birthYear, Account account) {
+    public PersonTag(String name, short birthYear, Account account) {
         this.name = name;
         this.birthYear = birthYear;
         this.account = account;
